#29 [Assert] All-in Hamcrest All custom verbose performers are removed All excessive methods are removed from the AssertBuilder and the Assert
