--- src/main/java/org/whaka/asserts/ThrowableAssert.java
@@ -1,32 +1,0 @@
-package org.whaka.asserts;
-
-import org.whaka.asserts.builder.AssertBuilder;
-import org.whaka.asserts.builder.ThrowableAssertPerformer;
-
-public class ThrowableAssert extends AbstractInstantAssert<ThrowableAssertPerformer, Throwable> {
-
-	public ThrowableAssert(Throwable actual) {
-		super(actual);
-	}
-	
-	@Override
-	protected ThrowableAssertPerformer createPerformer(AssertBuilder builder, Throwable actual) {
-		return builder.checkThrowable(actual);
-	}
-	
-	public void notExpected() {
-		notExpected(null);
-	}
-	
-	public void notExpected(String message) {
-		performInstantAssert(ThrowableAssertPerformer::notExpected, message);
-	}
-
-	public void isInstanceOf(Class<? extends Throwable> type) {
-		isInstanceOf(type, null);
-	}
-	
-	public void isInstanceOf(Class<? extends Throwable> type, String message) {
-		performInstantAssert(throwablePerformer -> throwablePerformer.isInstanceOf(type), message);
-	}
-}
